fix(core): Move mountable's asset to scope of target
We're unable to update the mountable's scripts without a full tear-down of all consuming stacks.
This moves the assets to the target's stack instead of the Mountable's stack to avoid exports in
cross-stack usage.

BREAKING CHANGE: Stacks set up like our examples will see an error regarding being unable to
 delete an export in use when deploying a stack update. To bypass, use cdk deploy's -e option
 to deploy each stack upstream of the Mountable's stack before updating the Mountable's stack. e.g.
- cdk deploy -e ComputeTier; cdk deploy -e ServiceTier; cdk deploy -e StorageTier